improve www LCP/FCP: preload fonts, trim PostHog, lazy docs UI
The H1 LCP element on /www had ~790ms of element-render-delay almost
entirely from the HTML→CSS→font waterfall, and the main bundle still
included fumadocs-ui (RootProvider, SearchDialog), shiki, and orama
because routes/docs/$.tsx pulled them in at the module top level.

- Preload geist-sans, geist-mono, titan-one woff2 in __root head so
  fonts download in parallel with the CSS instead of after it.
- PostHog: disable_surveys, capture_dead_clicks: false,
  capture_performance: false. The server-side decide response already
  returns surveys: false etc., but the SDK was still proactively
  fetching /static/{surveys,dead-clicks-autocapture,web-vitals}.js.
- Move DocsPageLayout, clientLoader, DocsPageActions, OpenApiContent,
  and all their fumadocs-ui / docs-component imports into a new
  components/docs-page-layout.tsx. Both /docs/$ and /docs/ reach it
  via a static import inside the route's component, so the imports
  follow the auto-code-split into the docs chunk and stop loading on
  the marketing pages.
- defaultPreload: "intent" on the router so hovering a Docs link
  preloads the route component (and its lazy chunk) before click,
  replacing the dropped clientLoader.preload call.
- Dynamic source import inside the createServerFn handler avoids
  bundling fumadocs-core/source + fumadocs-openapi (with its
  top-level await openapiSource(...)) into the client.
