+++ apps/www/src/router.tsx
@@ -7,6 +7,10 @@ export function getRouter() {
 		routeTree,
 		scrollRestoration: true,
 		defaultNotFoundComponent: NotFound,
+		// Preload the matched route (component chunk + loader) on link hover so
+		// navigating from the marketing pages into /docs doesn't feel like a
+		// hard reload while the lazy fumadocs-ui bundle downloads.
+		defaultPreload: "intent",
 	});
 
 	return router;
